backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
requests
pandas
numpy
scikit-learn
joblib
